Give priority to libraries installed in the homebrew path on macOS
macOS comes with an ancient version of openssl that the tor autoconf
script will resolve to first. Instead, if the user has installed
homebrew, it should first pick the homebrew version first.

If the user specifies the `--with-$x-dir` `./configure` option, that
should still be respected.

This fix should not affect any other platform as the path we check to
see if exists is `/usr/local/Cellar` which is specific to homebrew.

More details about the problem this will solve can be found here:
https://tor.stackexchange.com/questions/595/tor-compiling-error-undefined-symbols-in-libor-crypto-a